zrpcd: update ethtag type to i64
In RFC 7432, EVPN RT1 ethtag field can take the MAX_ET=0xffffffff value
which can not be set with signed 32 bits integer.
Even if using the -1 value would do the job, it is not a robust choice
to ask user to set -1 in vpnservice.thrift interface for ethtag value,
wondering that the thrift underlying langage would interprete it as
0xffffffff.

+++ zrpcd/zrpc_bgp_configurator.c
@@ -46,12 +46,12 @@ instance_bgp_configurator_handler_unset_ebgp_multihop(BgpConfiguratorIf *iface, 
                                                       const gchar * peerIp, GError **error);
 gboolean
 instance_bgp_configurator_handler_push_route(BgpConfiguratorIf *iface, gint32* _return, const protocol_type p_type, const gchar * prefix,
-                                             const gchar * nexthop, const gchar * rd, const gint32 ethtag, const gchar * esi,
+                                             const gchar * nexthop, const gchar * rd, const gint64 ethtag, const gchar * esi,
                                              const gchar * macaddress, const gint32 l3label, const gint32 l2label,
                                              const encap_type enc_type, const gchar * routermac, GError **error);
 gboolean
 instance_bgp_configurator_handler_withdraw_route(BgpConfiguratorIf *iface, gint32* _return, const protocol_type p_type, const gchar * prefix,
-                                                 const gchar * rd,  const gint32 ethtag, const gchar * esi, const gchar * macaddress, GError **error);
+                                                 const gchar * rd,  const gint64 ethtag, const gchar * esi, const gchar * macaddress, GError **error);
 gboolean
 instance_bgp_configurator_handler_stop_bgp(BgpConfiguratorIf *iface, gint32* _return, const gint64 asNumber, GError **error);
 gboolean
@@ -817,7 +817,7 @@ instance_bgp_configurator_handler_unset_ebgp_multihop(BgpConfiguratorIf *iface, 
  */
 gboolean
 instance_bgp_configurator_handler_push_route(BgpConfiguratorIf *iface, gint32* _return, const protocol_type p_type, const gchar * prefix,
-                                             const gchar * nexthop, const gchar * rd, const gint32 ethtag, const gchar * esi,
+                                             const gchar * nexthop, const gchar * rd, const gint64 ethtag, const gchar * esi,
                                              const gchar * macaddress, const gint32 l3label, const gint32 l2label, 
                                              const encap_type enc_type, const gchar * routermac, GError **error)
 {
@@ -952,7 +952,7 @@ instance_bgp_configurator_handler_push_route(BgpConfiguratorIf *iface, gint32* _
  */
 gboolean
 instance_bgp_configurator_handler_withdraw_route(BgpConfiguratorIf *iface, gint32* _return, const protocol_type p_type, const gchar * prefix,
-                                                 const gchar * rd,  const gint32 ethtag, const gchar * esi, const gchar * macaddress, GError **error)
+                                                 const gchar * rd,  const gint64 ethtag, const gchar * esi, const gchar * macaddress, GError **error)
 {
   struct zrpc_vpnservice *ctxt = NULL;
   struct bgp_api_route inst;
